# docs/openapi-schema.yml
---
openapi: 3.1.0
info:
  title: Restaurant Menu API
  description: REST API по работе с меню ресторана.
  version: 0.1.0
paths:
  "/api/v1/menus":
    get:
      tags:
      - Меню
      summary: Список меню
      description: Выводим список со всеми меню.
      operationId: all_menus_api_v1_menus_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/DetailedMenuInfoPyd"
                type: array
                title: Response All Menus Api V1 Menus Get
    post:
      tags:
      - Меню
      summary: Создать меню
      description: Создаём новое меню.
      operationId: new_menu_api_v1_menus_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BaseMenuPyd"
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedMenuInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/menus/{menu_id}":
    get:
      tags:
      - Меню
      summary: Определённое меню
      description: Выводим определённое меню по его «id».
      operationId: get_menu_api_v1_menus__menu_id__get
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedMenuInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    patch:
      tags:
      - Меню
      summary: Обновить меню
      description: Обновляем информацию о меню.
      operationId: update_menu_api_v1_menus__menu_id__patch
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateMenuPyd"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedMenuInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      tags:
      - Меню
      summary: Удалить меню
      description: Удалаяем меню.
      operationId: delete_menu_api_v1_menus__menu_id__delete
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DeleteObjPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/menus/{menu_id}/submenus":
    post:
      tags:
      - Подменю
      summary: Создать подменю
      description: Создаём новое подменю, для определённого меню.
      operationId: new_submenu_api_v1_menus__menu_id__submenus_post
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BaseMenuPyd"
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedSubmenuInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    get:
      tags:
      - Подменю
      summary: Список подменю
      description: Выводим список со всеми подменю, для определённого меню.
      operationId: all_submenus_api_v1_menus__menu_id__submenus_get
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/DetailedSubmenuInfoPyd"
                title: Response All Submenus Api V1 Menus  Menu Id  Submenus Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/menus/{menu_id}/submenus/{submenu_id}":
    get:
      tags:
      - Подменю
      summary: Определённое подменю
      description: Выводим определённое подменю.
      operationId: get_submenu_api_v1_menus__menu_id__submenus__submenu_id__get
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedSubmenuInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    patch:
      tags:
      - Подменю
      summary: Обновить подменю
      description: Обновляем информацию о подменю.
      operationId: update_submenu_api_v1_menus__menu_id__submenus__submenu_id__patch
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateMenuPyd"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedSubmenuInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      tags:
      - Подменю
      summary: Удалить подменю
      description: Удалаяем подменю.
      operationId: delete_submenu_api_v1_menus__menu_id__submenus__submenu_id__delete
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DeleteObjPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes":
    post:
      tags:
      - Блюдо
      summary: Создать блюдо
      description: Создаём новое блюдо, для определённого подменю.
      operationId: new_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes_post
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BaseDishPyd"
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedDishInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    get:
      tags:
      - Блюдо
      summary: Список блюд
      description: Выводим список со всеми блюдами, для определённого подменю.
      operationId: all_dishes_api_v1_menus__menu_id__submenus__submenu_id__dishes_get
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/DetailedDishInfoPyd"
                title: Response All Dishes Api V1 Menus  Menu Id  Submenus  Submenu
                  Id  Dishes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes/{dish_id}":
    get:
      tags:
      - Блюдо
      summary: Определённое блюдо
      description: Выводим определённое блюдо.
      operationId: get_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__get
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      - name: dish_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id блюда
          title: Dish Id
        description: id блюда
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedDishInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    patch:
      tags:
      - Блюдо
      summary: Обновить блюдо
      description: Обновляем информацию о блюде.
      operationId: update_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__patch
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      - name: dish_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id блюда
          title: Dish Id
        description: id блюда
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateDishPyd"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedDishInfoPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
    delete:
      tags:
      - Блюдо
      summary: Удалить блюдо
      description: Удалаяем блюдо.
      operationId: delete_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__delete
      parameters:
      - name: menu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id меню
          title: Menu Id
        description: id меню
      - name: submenu_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id подменю
          title: Submenu Id
        description: id подменю
      - name: dish_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: id блюда
          title: Dish Id
        description: id блюда
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DeleteObjPyd"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
components:
  schemas:
    BaseDishPyd:
      properties:
        title:
          type: string
          title: Title
          description: Название блюда
        description:
          type: string
          title: Description
          description: Описание блюда
        price:
          type: number
          title: Price
          description: Цена блюда
      type: object
      required:
      - title
      - description
      - price
      title: BaseDishPyd
      description: |-
        Pydantic модель c базовой информацией о блюде.

        Fields:
            - title: str
            - description: str
            - price: float
    BaseMenuPyd:
      properties:
        title:
          type: string
          title: Title
          description: Название меню
        description:
          type: string
          title: Description
          description: Описание меню
      type: object
      required:
      - title
      - description
      title: BaseMenuPyd
      description: |-
        Pydantic модель c базовой информацией о меню/подменю.

        Fields:
            - title: str
            - description: str
    DeleteObjPyd:
      properties:
        status:
          type: boolean
          title: Status
          description: Статус удаления.
        message:
          type: string
          title: Message
          description: Сообщение, с описанием результата.
      type: object
      required:
      - status
      - message
      title: DeleteObjPyd
      description: |-
        Pydantic модель для удаления меню/подменю/блюда.

        Fields:
            - status: bool
            - message: str
    DetailedDishInfoPyd:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: id блюда в БД
        title:
          type: string
          title: Title
          description: Название блюда
        description:
          type: string
          title: Description
          description: Описание блюда
        price:
          anyOf:
          - type: number
          - type: string
          title: Price
          description: Цена блюда
      type: object
      required:
      - id
      - title
      - description
      - price
      title: DetailedDishInfoPyd
      description: |-
        Pydantic модель с подробной информацией о блюде.

        Fields:
            - id: UUID
            - title: str
            - description: str
            - price: Union[float, str]
    DetailedMenuInfoPyd:
      properties:
        title:
          type: string
          title: Title
          description: Название меню
        description:
          type: string
          title: Description
          description: Описание меню
        id:
          type: string
          format: uuid
          title: Id
          description: id меню в БД
        submenus_count:
          type: integer
          title: Submenus Count
          description: Количество подменю
        dishes_count:
          type: integer
          title: Dishes Count
          description: Количество блюд
      type: object
      required:
      - title
      - description
      - id
      - submenus_count
      - dishes_count
      title: DetailedMenuInfoPyd
      description: |-
        Pydantic модель с подробной информацией о меню.

        Fields:
            - id: UUID
            - title: str
            - description: str
            - submenus_count: int
            - dishes_count: int
    DetailedSubmenuInfoPyd:
      properties:
        title:
          type: string
          title: Title
          description: Название меню
        description:
          type: string
          title: Description
          description: Описание меню
        id:
          type: string
          format: uuid
          title: Id
          description: id подменю в БД
        dishes_count:
          type: integer
          title: Dishes Count
          description: Количество блюд
      type: object
      required:
      - title
      - description
      - id
      - dishes_count
      title: DetailedSubmenuInfoPyd
      description: |-
        Pydantic модель с подробной информацией о подменю.

        Fields:
            - id: UUID
            - title: str
            - description: str
            - dishes_count: int
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    UpdateDishPyd:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          description: Название блюда
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Описание блюда
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
          description: Цена блюда
      type: object
      title: UpdateDishPyd
      description: |-
        Pydantic модель для обновления информации о блюде.

        Fields:
            - title: str | None
            - description: str | None
            - price: float | None
    UpdateMenuPyd:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          description: Название меню
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Описание меню
      type: object
      title: UpdateMenuPyd
      description: |-
        Pydantic модель для обновления меню/подменю.

        Fields:
            - title: str | None
            - description: str | None
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError